moved HolonId byte encoding helpers to ids.rs.  removed the requirement for descriptors to be supplied with dance input parameters.
